moving artifact related code into helper file
